requirements.txt
fastapi~=0.111.0
SQLAlchemy~=2.0.31
DateTime~=5.5
pydantic~=2.7.4
requests~=2.32.0
tqdm~=4.66.5
python-dotenv~=1.0.1
pandas~=2.1.1
geopy~=2.4.1